+++ src/ex/arr_bug.bu
