+++ src/new/mod.rs
@@ -1,5 +1,4 @@
 pub mod nodes;
 pub mod instr;
-pub mod new_codegen;
 pub mod new_assembler;
 pub mod compiler;
